refactor: move compound out of searchers
