+++ jinahub/indexers/compound/FaissLMDBSearcher/Dockerfile
