+++ jinahub/indexers/compound/FaissLMDBSearcher/config.yml
