+++ jinahub/indexers/compound/FaissLMDBSearcher/manifest.yml
