+++ jinahub/indexers/compound/FaissPostgresIndexer/Dockerfile
