+++ jinahub/indexers/compound/FaissPostgresIndexer/manifest.yml
