+++ jinahub/indexers/compound/FaissPostgresIndexer/tests/docker_args.txt
