+++ jinahub/indexers/compound/FaissPostgresIndexer/tests/integration/test_dummy.py
