+++ jinahub/indexers/compound/FaissPostgresIndexer/tests/requirements.txt
