+++ tests/integration/faiss_psql/test_flow_integration.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from jina import Document, Flow
 
-from jinahub.indexers.searcher.compound.FaissPostgresIndexer import FaissPostgresIndexer
+from jinahub.indexers.compound.FaissPostgresIndexer import FaissPostgresIndexer
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 compose_yml = os.path.join(cur_dir, 'docker-compose.yml')
